# This is a combination of 3 commits. # The first commit's message is: Switched to Gradle
# This is the 2nd commit message:

Travis Support

# This is the 3rd commit message:

Remove Classpath
