--- .classpath
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="var" path="wpilib" sourcepath="wpilib.sources"/>
-	<classpathentry kind="var" path="networktables" sourcepath="networktables.sources"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/Team1389-Robot-Code-Base"/>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/gson-2.3.1.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/javax.servlet-api-3.1.0.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/jetty-http-9.3.2.v20150730.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/jetty-io-9.3.2.v20150730.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/jetty-security-9.3.2.v20150730.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/jetty-server-9.3.2.v20150730.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/jetty-servlet-9.3.2.v20150730.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/jetty-util-9.3.2.v20150730.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/strongback.jar" sourcepath="/Team1389-Robot-Code-Base/lib/strongback-sources.jar"/>
-	<classpathentry kind="lib" path="/Team1389-Robot-Code-Base/lib/strongback-testing.jar" sourcepath="/Team1389-Robot-Code-Base/lib/strongback-testing-sources.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
