+++ .travis.yml
@@ -1,0 +1,13 @@
+sudo: required
+dist: trusty
+
+language: java
+jdk:
+  - openjdk8
+  - oraclejdk8
+
+env: GRADLE_OPTS="-Xms1g -Xmx3g"
+script: ./gradlew build
+
+notifications:
+  email: false
