+++ README.md
@@ -1,4 +1,3 @@
 This is our code for our 2016-bot
 
-To use this repository, make sure to clone the Team1389-Robot-Base-Code repo,
-as well as follow instruction in the base.properties file in this repo
+To use this repository, make sure to clone the Team1389-Robot-Base-Code repo.
