--- base.properties
@@ -1,3 +1,0 @@
-#set this value to the location of the Team1389-Base-Project-Template (NOTE: make sure to use forward slashes)
-#this will be different on every computer
-team1389base.loc=/home/user/git/Team1389-Robot-Code-Base
