--- build.properties
@@ -1,11 +1,0 @@
-# Project specific information
-package=com.team1389.y2016.robot
-robot.class=${package}.HardwareMain
-simulation.world.file=/usr/share/frcsim/worlds/GearsBotDemo.world
-
-team1389base.jar = ${team1389base.loc}/export/Team1389-Robot-Code-Base.jar
-
-project.resources.jar = build/jars/projectResources.jar
-project.resources = ${src.dir}/../Resources
-userLibs=${team1389base.loc}/export/Team1389-Robot-Code-Base.jar
-#userLibs=${team1389base.jar}
