+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Feb 07 17:02:06 PST 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.5-rc-2-all.zip
